requirements.txt
fastapi
uvicorn
requests
geopy
pydantic
httpx
pytest
hypothesis
pytest-mock
python-dotenv
